apps/functions-archive/functions/requirements.txt
# Requires Python >= 3.11
# Azure Functions runtime
azure-functions==1.17.0

# Azure SDK
azure-core==1.35.0
azure-cosmos==4.5.1
# Note: Azure Computer Vision - Image Analysis API will be retired on September 25, 2028
# Migration recommended by September 2026. See docs/azure-vision-migration.md for tracking.
azure-ai-vision-imageanalysis==1.0.0

# OpenAI
openai==1.35.13

# Email
# Note: smtplib is built-in, no need to install

# Stripe
stripe==10.7.0

# Additional utilities
python-dotenv==1.0.0
email-validator==2.1.0
